Algorithm/ZenHealthAppEngine/requirements.txt
Flask==0.12.2
gunicorn==19.7.1
scikit-learn==0.19.1
matplotlib==2.1.0
scipy==1.0.0
numpy==1.13.3
pandas==0.21.0
pip==10.0.1
python-firebase==1.2
#google-cloud-storage==1.2.0
